Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/xhr/event-timeout-order.any-expected.txt


FAIL XMLHttpRequest: event - timeout (order of events) assert_equals: expected "readystatechange(4)" but got "upload.progress(12,12,true)"
